Feat/pys

+++ src/main/java/com/okestro/community/service/BoardService.java
@@ -8,6 +8,6 @@ public interface BoardService {
 
     List<ListDto> boardlist();
 
-    ListDto getBoardById(Long id);
+    ListDto getBoardById(int id);
 
 }
